Modularized database
