+++ db/schema.js
@@ -1,0 +1,16 @@
+const db = require('./index.js');
+const mongoose = require('mongoose');
+
+const productSchema = mongoose.Schema({
+  productName: {
+    type: String,
+    unique: true,
+    index: true
+  },
+  bannerImageUrl: String,
+  productImageUrls: String
+}, { strict: true });
+
+const Product = mongoose.model('Product', productSchema);
+
+exports.Product = Product;
